Refactor IT tests
