+++ mockguard/src/test/kotlin/com/mockguard/integration/fixtures/MockGuardTestDoubles.kt
@@ -1,0 +1,9 @@
+package com.mockguard.integration.fixtures
+
+interface Dependency {
+    fun call()
+}
+
+interface Logger {
+    fun info(message: String)
+}
